Move eth2 withdrawals test to the proper location
